bill.ts: improve display of additional information
According to the SIX implementation guide, additional information can contain
both unstructured message and bill information.

They can be separated by a new line and the total length must be of 140 caracters.

To be able to keep track of the newY position I had to introduce a little hack
which split message on 40-length chunks.