automatic text wrap and truncate
The guideline recommends to split creditor and debtor name on two lines
if it's too large according the the section width. If it's still too
large, we should truncate.

There's no recommendation according to the creditor and debtor address.
So we split one two lines too and truncate if it's too big.
This is an improvement, because otherwise address will be printed above
next section.

For the additional information section, it only set maximum size with
140 characters and add an ellipsis if it's bigger.